api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.38.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiosmtplib==3.0.0
python-dotenv==1.0.0
httpx==0.25.0